Always ask lhs instead of choosing one side at random.
If the user wants to be asked both ways, they can simply copy an entry
and swap the sides of the copied entry.

Rationale:
Whether the user be asked one side or the other should be up to the
user themself. Else there's no way of making the program ask one side
only, which definitely is a use-case.